+++ tests/testthat/test-Contours.R
@@ -48,6 +48,32 @@ test_that('Contours are plotted', {
   }
   expect_doppelganger('Contours', Contours)
   
+  ContoursSkiwiff <- function () {
+    FunctionToContour <- function (a, b, c) {
+      a - c + (4 * a * b) + (27 * a * b * c)
+    }
+    SubTest <- function (direction) {
+      ColourTernary(TernaryPointValues(FunctionToContour, resolution=6L, 
+                                       direction = direction))
+      TernaryContour(FunctionToContour, resolution = 12L, direction = direction)
+    }
+    
+    par(mar=rep(0, 4), mfrow=c(2, 2))
+    TernaryPlot(point = 3L, ylim=c(0, 1))
+    SubTest(1)
+    
+    TernaryPlot(point = 4L, xlim=c(0, 1))
+    SubTest(2)
+    
+    TernaryPlot(point = 1L, ylim=c(-1, 0))
+    SubTest(3)
+    
+    TernaryPlot(point = 2L, xlim=c(-1, 0))
+    SubTest(4)
+  }
+  expect_doppelganger('Contours-skiwiff', ContoursSkiwiff)
+  
+  
   DensityContours <- function () {
     par(mar=rep(0.2, 4))
     TernaryPlot()