+++ NEWS.md
@@ -1,10 +1,13 @@
+# Ternary v1.1.3
+- Use `inherits`, not `class` [https://developer.r-project.org/Blog/public/2019/11/09/when-you-think-class.-think-again].
+- Obey user-specified `direction` in `TernaryContour` and `TernaryPointValues`
+  (thanks @pkR-pjR).
+
 # Ternary v1.1.2
-## New features
  - `clockwise` parameter allows plots to be configured in 'flipped' orientation
    (https://github.com/ms609/Ternary/issues/22)
 
 # Ternary v1.1.1
-## New features
  - `TernaryArrows` adds arrows to a ternary plot.
 
 # Ternary v1.1.0